Apply further abstractions for memory and board functions
