+++ CMakeLists.txt
@@ -34,6 +34,7 @@ INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/arch/${CMAKE_SYSTEM_PROCESSOR})
 INCLUDE_DIRECTORIES(Common)
 INCLUDE_DIRECTORIES(System)
 INCLUDE_DIRECTORIES(BoardSupport/Common)
+INCLUDE_DIRECTORIES(${TARGET_SYSTEM_BOARD})
 
 
 IF (${FSANITIZE})
